+++ src/redux/reducers/winterAreaReducers.ts
